refactor: created separate folder for Navbar
