+++ src/Layout.jsx
@@ -1,4 +1,4 @@
-import Navbar from "./components/Navbar";
+import Navbar from "./components/Navbar/Navbar";
 import { useLocation } from "react-router-dom";
 
 /*
