--- src/components/buttons/Dropdown.css
@@ -1,43 +1,0 @@
-.scrollable-dropdown {
-    max-height: 250px;
-    max-width: 200px;
-    overflow-y: auto;
-}
-
-.scrollable-dropdown .dropdown-item {
-    white-space: normal;
-    word-wrap: break-word;
-    border-top: 1px solid #D6D6D6;
-}
-
-.dropdown-toggle {
-    background-color: white;
-}
-
-.custom-dropdown-toggle {
-    background: none !important;
-    border: none !important;
-    color: #1a1a19 !important;
-    padding: 0.5rem !important;
-    margin: 0 0.5rem !important;
-    border-radius: 0.5rem !important;
-    font-weight: 500 !important;
-    cursor: pointer !important;
-}
-
-.custom-dropdown-toggle:hover {
-    background: none !important;
-    border: none !important;
-    color: #FF9843 !important;
-    padding: 0.5rem !important;
-    margin: 0 0.5rem !important;
-    border-radius: 0.5rem !important;
-    font-weight: 500 !important;
-    cursor: pointer !important;
-}
-
-.custom-dropdown-toggle:focus {
-    outline: none;
-    box-shadow: none;
-    /* Removes the blue focus glow */
-}
