+++ src/components/ui/button.jsx
@@ -18,7 +18,8 @@ const buttonVariants = cva(
           "border border-input bg-background hover:bg-accent hover:text-accent-foreground",
         secondary:
           "bg-secondary text-secondary-foreground hover:bg-secondary/80",
-        ghost: "hover:bg-accent hover:text-accent-foreground",
+        ghost:
+          "hover:bg-accent hover:text-accent-foreground focus-visible:outline-none",
         link: "text-primary underline-offset-4 hover:underline",
         success: "bg-green-600 text-primary-foreground hover:bg-green-600/90",
         active:
